added styling for main page
