+++ package.json
@@ -7,7 +7,8 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node server.js",
     "local": "./node_modules/.bin/nodemon -L --ignore public/",
-    "db:reset": "node bin/resetdb.js"
+    "db:reset": "node bin/resetdb.js",
+    "sass-watch": "sass --watch styles:public/styles --poll --no-source-map"
   },
   "author": "Lighthouse Labs",
   "license": "ISC",
