+++ styles/main.scss
@@ -1,7 +1,4 @@
 body {
-  display: flex;
-  justify-content: center;
-  align-items: center;
   min-height: 100vh;
 
   background: rgb(163, 208, 219);
