+++ styles/quiz.scss
@@ -1,0 +1,9 @@
+.quiz {
+  &_my_quizzes {
+    margin: 5rem 5rem;
+  }
+  &_card {
+    border: 2px solid black;
+    width: 50%;
+  }
+}
